+++ staff.rb
@@ -1,0 +1,51 @@
+require_relative 'people'
+
+class Staff < People
+	attr_reader :age, :salary, :phase, :target_raise
+  attr_accessor :name
+
+  def initialize(options={})
+    # @age = options.fetch(:age, 0)
+    # @name = options.fetch(:name, "")
+    super
+    @phase = 3
+  end
+
+  # def offer_high_five
+  #   "High five!"
+  # end
+
+  def set_phase(num)
+    @phase = num
+    "Cool, I've always wanted to teach phase #{num}!"
+  end
+
+  def teach_stuff
+    response = ""
+    response += "Listen, class, this is how everything works. "
+    response += "*drops crazy knowledge bomb* "
+    response += "... You're welcome."
+    response
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+  def set_performance_rating(rating)
+    response = ""
+    if rating > self.class::RATING
+      receive_raise(@target_raise)
+      response = "Yay, I'm a great employee!"
+    else
+      response += "Oh, well -- thanks to this actionable, specific, and kind "
+      response += self.class::RESPONSE
+    end
+    response
+  end
+end
