+++ student.rb
@@ -1,16 +1,19 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'people'
+
+class Student < People
+  # attr_reader :age, :phase
+  # attr_accessor :name
 
   def initialize(options = {})
+    # @age = options.fetch(:age, 0)
+    # @name = options.fetch(:name, "")
+    super
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
 
   def set_phase(num)
     response = ""
