+++ teacher.rb
@@ -1,49 +1,55 @@
-class Teacher
-  attr_reader :age, :salary, :phase, :performance_rating, :target_raise
-  attr_accessor :name
+require_relative 'staff'
 
-  def initialize(options={})
-    @phase = 3
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+class Teacher < Staff
+
+  RATING = 90
+  RESPONSE = "feedback, I'll do better next time."
+  # attr_reader :age, :salary, :phase, :performance_rating, :target_raise
+  # attr_accessor :name
+
+   def initialize(options={})
+  #   @age = options.fetch(:age, 0)
+  #   @name = options.fetch(:name, "")
+   super
     @target_raise = 1000
-  end
+    # @phase = 3
+   end
 
-  def offer_high_five
-    "High five!"
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
 
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
+  # def set_phase(num)
+  #   @phase = num
+  #   "Cool, I've always wanted to teach phase #{num}!"
+  # end
 
-  def teach_stuff
-    response = ""
-    response += "Listen, class, this is how everything works, fo SHO! "
-    response += "*drops flat-out insane knowledge bomb* "
-    response += "... You're welcome. *saunters away*"
-    response
-  end
+  # def teach_stuff
+  #   response = ""
+  #   response += "Listen, class, this is how everything works, fo SHO! "
+  #   response += "*drops flat-out insane knowledge bomb* "
+  #   response += "... You're welcome. *saunters away*"
+  #   response
+  # end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
+  # def salary=(new_salary)
+  #   puts "This better be good!"
+  #   @salary = new_salary
+  # end
 
-  def receive_raise(raise)
-    @salary += raise
-  end
+  # def receive_raise(raise)
+  #   @salary += raise
+  # end
 
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      receive_raise(@target_raise)
-      response = "Yay, I'm a great employee!"
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
-  end
-end
+  # # def set_performance_rating(rating)
+  # #   response = ""
+  #   if rating > 90
+  #     receive_raise(@target_raise)
+  #     response = "Yay, I'm a great employee!"
+  #   else
+  #     response += "Oh, well -- thanks to this actionable, specific, and kind "
+  #     response += "feedback, I'll do better next time."
+  #   end
+  #   response
+  # end
+   end
